Include check for args in suggest command
Also fixes message.author.name >>> message.member.displayName

+++ commands/suggest.js
@@ -8,20 +8,21 @@ module.exports = {
         if (!args) return message.delete();
 
         let messageArgs = args.join(' ');
-        const embed = new Discord.MessageEmbed()
-            .setColor('#ff3300')
-            .setAuthor(message.author.name, message.author.displayAvatarURL({dynamic: true}))
-            .setTitle('New suggestion:')
-            .setDescription(messageArgs);
+        if (args) {
+            const embed = new Discord.MessageEmbed()
+                .setColor('#ff3300')
+                .setAuthor(message.member.displayName, message.author.displayAvatarURL({dynamic: true}))
+                .setTitle('New suggestion:')
+                .setDescription(messageArgs);
 
-        channel.send(embed)
-            .then((msg) => {
-                msg.react('👍');
-                msg.react('👎');
-                message.delete();
-            }).catch((err) => {
-            throw err;
-        });
-
+            channel.send(embed)
+                .then((msg) => {
+                    msg.react('👍');
+                    msg.react('👎');
+                    message.delete();
+                }).catch((err) => {
+                throw err;
+            });
+        }
     }
 };
